Convert to Maven project
After many hours of blood, sweat, tears and frustration - I am proud to say that the project now compiles correctly with Maven and is runnable using jetty-runner.

There really was no reason for this, the project built and ran correctly using the IDEs built-in tools, however what would be the fun in making things easy for myself, huh?